requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
requests
python-multipart
psd_tools
reportlab